+++ samcli/commands/check/resources/graph.py
