+++ samcli/commands/check/resources/lambda_function_pricing.py
