+++ tests/unit/commands/check/test_pricing.py
@@ -1,13 +1,13 @@
-from samcli.commands.check.resources.LambdaFunctionPricing import LambdaFunctionPricing
+from samcli.commands.check.resources.lambda_function_pricing import LambdaFunctionPricing
 from unittest import TestCase
 from unittest.mock import Mock, patch
 
-from samcli.commands.check.resources.Pricing import Pricing, _ask
+from samcli.commands.check.resources.pricing import Pricing, _ask
 from samcli.commands._utils.resources import AWS_LAMBDA_FUNCTION
 
 
 class TestPricing(TestCase):
-    @patch("samcli.commands.check.resources.Pricing.click")
+    @patch("samcli.commands.check.resources.pricing.click")
     def test_ask(self, patch_click):
         question_mock = "Mock()"
 
@@ -18,7 +18,7 @@ class TestPricing(TestCase):
         patch_click.prompt.assert_called_once_with(text=question_mock, type=int)
         self.assertEqual(result, patch_click.prompt.return_value)
 
-    @patch("samcli.commands.check.resources.Pricing.click")
+    @patch("samcli.commands.check.resources.pricing.click")
     def test_ask_memory(self, patch_click):
         graph_mock = Mock()
         question_mock = "Mock()"
@@ -39,7 +39,7 @@ class TestPricing(TestCase):
         self.assertEqual(result0, split_value[0])
         self.assertEqual(result1, split_value[1])
 
-    @patch("samcli.commands.check.resources.Pricing.click")
+    @patch("samcli.commands.check.resources.pricing.click")
     def test_correct_memory_input(self, patch_click):
         self_mock = Mock()
         self_mock._min_memory_amount = 128
@@ -121,8 +121,8 @@ class TestPricing(TestCase):
 
         self_mock._ask_lambda_function_questions.assert_called_once()
 
-    @patch("samcli.commands.check.resources.Pricing._ask")
-    @patch("samcli.commands.check.resources.Pricing.LambdaFunctionPricing")
+    @patch("samcli.commands.check.resources.pricing._ask")
+    @patch("samcli.commands.check.resources.pricing.LambdaFunctionPricing")
     def test_ask_lambda_function_questions(self, patch_LFPricing, patch_ask):
         pricing_instance_mock = Mock()
         patch_LFPricing.return_value = pricing_instance_mock
